Made Corrections - Renamed protocols by conforming to the Swift best practices - Renamed the Cells XIBs and Swift files to conform to the Swift best practices - Renamed the MARK headers to have shorter meaningful keywords
