+++ Lesson4/Controllers/MainTabBarController.swift
@@ -11,13 +11,13 @@ class MainTabBarController: UITabBarController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        
         // set default tab bar color
         UITabBar.appearance().unselectedItemTintColor = .white.withAlphaComponent(0.6)
     }
     
     override func tabBar(_ tabBar: UITabBar, didSelect item: UITabBarItem) {
-        let index = tabBar.items?.firstIndex(of: item)
-        if let index = index {
+        if let index = tabBar.items?.firstIndex(of: item) {
             if index == 0 {
                 title = "Your News feeds"
                 navigationController?.navigationBar.isHidden = false
