+++ Lesson4/Models/Protocols/UserProtocol.swift
@@ -1,5 +1,5 @@
 //
-//  UserIdentifiable.swift
+//  UserProtocol.swift
 //  Lesson4
 //
 //  Created by Evans Owamoyo on 13.10.2021.
@@ -8,6 +8,6 @@
 import Foundation
 
 /// a protocol that recognises a user property
-protocol UserIdentifiable {
+protocol UserProtocol {
     var user: User? { get set }
 }
