+++ Lesson4/Views/Cells/ImageFeedCell.swift
@@ -1,8 +1,9 @@
+
 import UIKit
 
-class FeedCell_Content: FeedCell {
+class ImageFeedCell: FeedCell {
     override func loadData(with feed: Feed) {
         super.loadData(with: feed)
-        contentLabel.text = feed.content
+        feedImage?.image = feed.image
     }
 }
