+++ Lesson4/Views/Extensions/ReplaceSegue.swift
@@ -1,5 +1,5 @@
 //
-//  UIStoryboardSegue+Classes.swift
+//  ReplaceSegue.swift
 //  Lesson4
 //
 //  Created by Evans Owamoyo on 13.10.2021.
